# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bessel API",
    "summary": "Bessel HTTP and Webhooks API",
    "description": "Hello! This is the Bessel API.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://api.getbessel.com",
      "description": "Production environment",
      "x-speakeasy-server-id": "production"
    }
  ],
  "paths": {
...
    "/v1/transactions": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "List Transactions",
        "description": "List transactions.",
        "operationId": "list_transactions_v1_transactions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bank_account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by bank account ID(s).",
              "title": "Bank Account Id"
            },
            "description": "Filter by bank account ID(s)."
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category ID(s).",
              "title": "Category Id"
            },
            "description": "Filter by category ID(s)."
          },
          {
            "name": "uncategorized",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, only show transactions without a category.",
              "default": false,
              "title": "Uncategorized"
            },
            "description": "If true, only show transactions without a category."
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by direction: 'debit' or 'credit'.",
              "title": "Direction"
            },
            "description": "Filter by direction: 'debit' or 'credit'."
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in description (case-insensitive).",
              "title": "Search"
            },
            "description": "Search in description (case-insensitive)."
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (inclusive).",
              "title": "Date From"
            },
            "description": "Start date (inclusive)."
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (inclusive).",
              "title": "Date To"
            },
            "description": "End date (inclusive)."
          },
          {
            "name": "is_business",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by business flag.",
              "title": "Is Business"
            },
            "description": "Filter by business flag."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Page number, defaults to 1.",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number, defaults to 1."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Size of a page, defaults to 10. Maximum is 100.",
              "default": 10,
              "title": "Limit"
            },
            "description": "Size of a page, defaults to 10. Maximum is 100."
          },
          {
            "name": "sorting",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionSortProperty"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sorting criterion. Several criteria can be used simultaneously and will be applied in order. Add a minus sign `-` before the criteria name to sort by descending order.",
              "default": [
                "-transaction_date"
              ],
              "title": "Sorting"
            },
            "description": "Sorting criterion. Several criteria can be used simultaneously and will be applied in order. Add a minus sign `-` before the criteria name to sort by descending order."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "transactions"
        ],
        "summary": "Delete Transactions",
        "description": "Delete transactions by IDs.",
        "operationId": "delete_transactions_v1_transactions_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/import": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Import Transactions",
        "description": "Import transactions from a bank export (CSV or XLSX).\n\nDuplicate transactions (by dedup hash) are automatically skipped.",
        "operationId": "import_transactions_v1_transactions_import_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bank",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bank identifier, e.g. 'marginalen'.",
              "title": "Bank"
            },
            "description": "Bank identifier, e.g. 'marginalen'."
          },
          {
            "name": "bank_account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Bank account to attach transactions to.",
              "title": "Bank Account Id"
            },
            "description": "Bank account to attach transactions to."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_transactions_v1_transactions_import_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/bulk": {
      "patch": {
        "tags": [
          "transactions"
        ],
        "summary": "Bulk Update Transactions",
        "description": "Update category for a list of transactions by ID.",
        "operationId": "bulk_update_transactions_v1_transactions_bulk_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/transactions/{transaction_id}": {
      "patch": {
        "tags": [
          "transactions"
        ],
        "summary": "Update Transaction",
        "description": "Update a transaction.",
        "operationId": "update_transaction_v1_transactions__transaction_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/categorize-by-description": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Bulk Categorize by Description",
        "description": "Set category for all transactions matching the given description.",
        "operationId": "categorize_by_description_v1_transactions_categorize_by_description_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCategorizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkCategorizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/transactions/spending-by-category": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Monthly Spending by Category",
        "description": "Aggregate debit spending per category for a given month.",
        "operationId": "spending_by_category_v1_transactions_spending_by_category_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Year to query.",
              "title": "Year"
            },
            "description": "Year to query."
          },
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Month to query (1-12).",
              "title": "Month"
            },
            "description": "Month to query (1-12)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlySpendingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/monthly-flow": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Monthly Income & Expenses",
        "description": "Return income and expenses aggregated per month.",
        "operationId": "monthly_flow_v1_transactions_monthly_flow_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of months to look back (including current).",
              "default": 6,
              "title": "Months"
            },
            "description": "Number of months to look back (including current)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyFlowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/klarna/transactions": {
      "get": {
        "tags": [
          "klarna"
        ],
        "summary": "Get Klarna Transactions",
        "description": "Fetch Klarna transactions and return raw GraphQL response.",
        "operationId": "get_klarna_transactions_v1_klarna_transactions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Klarna Bearer token",
              "title": "Authorization"
            },
            "description": "Klarna Bearer token"
          },
          {
            "name": "cookie",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Klarna session cookies",
              "title": "Cookie"
            },
            "description": "Klarna session cookies"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Klarna Transactions V1 Klarna Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_transactions_v1_transactions_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "format": "binary"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_transactions_v1_transactions_import_post"
      },
...
      "BulkCategorizeResponse": {
        "properties": {
          "updated": {
            "type": "integer",
            "title": "Updated",
            "description": "Number of transactions updated."
          }
        },
        "type": "object",
        "required": [
          "updated"
        ],
        "title": "BulkCategorizeResponse"
      },
      "BulkDeleteRequest": {
        "properties": {
          "ids": {
            "items": {
              "type": "string",
              "format": "uuid4"
            },
            "type": "array",
            "title": "Ids",
            "description": "List of transaction IDs to delete."
          }
        },
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "BulkDeleteRequest"
      },
      "BulkUpdateRequest": {
        "properties": {
          "ids": {
            "items": {
              "type": "string",
              "format": "uuid4"
            },
            "type": "array",
            "title": "Ids",
            "description": "List of transaction IDs to update."
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid4"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id",
            "description": "Category ID to assign to all."
          }
        },
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "BulkUpdateRequest"
      },
      "BulkUpdateResponse": {
        "properties": {
          "updated": {
            "type": "integer",
            "title": "Updated",
            "description": "Number of transactions updated."
          }
        },
        "type": "object",
        "required": [
          "updated"
        ],
        "title": "BulkUpdateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportResponse": {
        "properties": {
          "created": {
            "type": "integer",
            "title": "Created",
            "description": "Number of new transactions imported."
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped",
            "description": "Number of duplicate transactions skipped."
          }
        },
        "type": "object",
        "required": [
          "created",
          "skipped"
        ],
        "title": "ImportResponse"
      },
...
      "TransactionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TransactionSchema"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "TransactionListResponse"
      },
...
      "TransactionSortProperty": {
        "type": "string",
        "enum": [
          "created_at",
          "-created_at",
          "transaction_date",
          "-transaction_date",
          "amount",
          "-amount",
          "description",
          "-description"
        ],
        "title": "TransactionSortProperty"
      },
      "TransactionUpdate": {
        "properties": {
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid4"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id",
            "description": "Category ID to assign."
          },
          "is_business": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Business",
            "description": "Mark as a business transaction."
          }
        },
        "type": "object",
        "title": "TransactionUpdate"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}